# src/main/java/ru/shaxowskiy/javaspeakerclub/dto/LectureCreateRequest.java
package ru.shaxowskiy.javaspeakerclub.dto;

import java.util.UUID;

public record LectureCreateRequest(
        String title,
        UUID talkId,
        Long speakerId,
        String mediaS3Key
) {}